Tidy up confusion among filters
Rename existing PredicateVersionFilter to ArtifactPredicateVersionFilter
to align it with existing VersionPredicateVersionFilter.

Also, UTs were in wrong package (typo).

Finally, fix snapshot and release filters, as they were wrong, they
totally neglected timestamped snapshots, and also dash in
"-SNAPSHOT" is not mandatory.
