+++ maven-resolver-util/src/test/java/org/eclipse/aether/util/graph/version/AbstractVersionFilterTest.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.eclipse.aether.util.graph.versions;
+package org.eclipse.aether.util.graph.version;
 
 import java.util.ArrayList;
 import java.util.Iterator;
