+++ maven-resolver-util/src/test/java/org/eclipse/aether/util/graph/version/PredicateVersionFilterTest.java
@@ -16,13 +16,12 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.eclipse.aether.util.graph.versions;
+package org.eclipse.aether.util.graph.version;
 
 import java.util.function.Predicate;
 
 import org.eclipse.aether.artifact.Artifact;
 import org.eclipse.aether.collection.VersionFilter.VersionFilterContext;
-import org.eclipse.aether.util.graph.version.PredicateVersionFilter;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
